requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
bs4
requests
